+++ src/main/java/shopping/dto/request/UpdateCartProductRequest.java
@@ -1,4 +1,4 @@
-package shopping.dto;
+package shopping.dto.request;
 
 public class UpdateCartProductRequest {
