+++ src/main/java/shopping/dto/response/FindCartProductResponse.java
@@ -1,4 +1,4 @@
-package shopping.dto;
+package shopping.dto.response;
 
 import shopping.domain.CartProduct;
