+++ src/main/java/shopping/dto/response/FindProductResponse.java
@@ -1,4 +1,4 @@
-package shopping.dto;
+package shopping.dto.response;
 
 import shopping.domain.Product;
