+++ src/main/java/shopping/ui/advice/GlobalControllerAdvice.java
@@ -4,7 +4,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
-import shopping.dto.ErrorResponse;
+import shopping.dto.response.ErrorResponse;
 import shopping.exception.AuthException;
 import shopping.exception.ShoppingException;
 import shopping.exception.TokenException;
