+++ src/test/java/shopping/application/CartProductServiceTest.java
@@ -10,9 +10,8 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import shopping.domain.CartProduct;
 import shopping.domain.Member;
 import shopping.domain.Product;
-import shopping.dto.FindCartProductResponse;
-import shopping.dto.UpdateCartProductRequest;
-import shopping.exception.CartException;
+import shopping.dto.response.FindCartProductResponse;
+import shopping.dto.request.UpdateCartProductRequest;
 import shopping.exception.MemberException;
 import shopping.exception.ProductException;
 import shopping.repository.CartProductRepository;
